+++ src/Windows Code Release 3/fileio.cpp
@@ -116,7 +116,7 @@ static bool FSClose(auto& file)
 	return !file.fail();
 }
 
-static bool SetFPos(auto& file, long len, std::ios_base::seekdir origin)
+static bool SetFPos(auto& file, size_t len, std::ios_base::seekdir origin)
 {
 	file.seekg(len, origin);
 	return !file.fail();
@@ -145,7 +145,7 @@ static T stream_read_type(std::istream& file)
 }
 #endif
 
-static void file_read_string(std::ifstream& file, long len, char* arr)
+static void file_read_string(std::ifstream& file, size_t len, char* arr)
 {
 	file.read(arr, len);
 	arr[len] = '\0';
@@ -433,9 +433,9 @@ void load_town(short town_num, short mode, short extra, char* str)
 
 	std::ifstream file_id;
 	short i, j;
-	long store;
+	size_t store;
 	bool success = false;
-	long len_to_jump = 0;
+	size_t len_to_jump = 0;
 	short which_town;
 
 	if (town_num != boe_clamp(town_num, 0, scenario.num_towns() - 1)) {
@@ -455,14 +455,14 @@ void load_town(short town_num, short mode, short extra, char* str)
 	len_to_jump = sizeof(scenario_data_type);
 	len_to_jump += sizeof(scen_item_data_type);
 	for (i = 0; i < 300; i++)
-		len_to_jump += (long)scenariodata.scen_str_len[i];
+		len_to_jump += (size_t)scenariodata.scen_str_len[i];
 	store = 0;
 	for (i = 0; i < 100; i++)
 		for (j = 0; j < 2; j++)
-			store += (long)(scenariodata.out_data_size[i][j]);
+			store += (size_t)(scenariodata.out_data_size[i][j]);
 	for (i = 0; i < which_town; i++)
 		for (j = 0; j < 5; j++)
-			store += (long)(scenariodata.town_data_size[i][j]);
+			store += (size_t)(scenariodata.town_data_size[i][j]);
 	len_to_jump += store;
 
 	if (!SetFPos(file_id, len_to_jump, std::ios_base::beg)) { FSClose(file_id); oops_error(35); }
@@ -550,7 +550,7 @@ void load_town(short town_num, short mode, short extra, char* str)
 	}
 
 	for (i = 0; i < 140; i++) {
-		const long len = (mode == 0) ? (long)(c_town.town.strlens[i]) : (long)(dummy_town.strlens[i]);
+		const size_t len = (mode == 0) ? (size_t)(c_town.town.strlens[i]) : (size_t)(dummy_town.strlens[i]);
 		switch (mode) {
 		case 0:
 			file_read_string(file_id, len, data_store.town_strs[i]);
@@ -579,7 +579,7 @@ void load_town(short town_num, short mode, short extra, char* str)
 		if (!success) { FSClose(file_id); oops_error(37); }
 
 		for (i = 0; i < 170; i++) {
-			file_read_string(file_id, (long)talking.strlens[i], data_store3.talk_strs[i]);
+			file_read_string(file_id, (size_t)talking.strlens[i], data_store3.talk_strs[i]);
 		}
 		cur_town_talk_loaded = town_num;
 	}
@@ -972,15 +972,15 @@ void load_outdoors(
 	}
 
 	const short out_sec_num = scenario.out_width() * to_create_y + to_create_x;
-	long len_to_jump = sizeof(scenario_data_type);
+	size_t len_to_jump = sizeof(scenario_data_type);
 	len_to_jump += sizeof(scen_item_data_type);
 	for (short i = 0; i < 300; i++)
-		len_to_jump += (long)scenariodata.scen_str_len[i];
+		len_to_jump += (size_t)scenariodata.scen_str_len[i];
 
-	long store = 0;
+	size_t store = 0;
 	for (short i = 0; i < out_sec_num; i++)
 		for (short j = 0; j < 2; j++)
-			store += (long)(scenariodata.out_data_size[i][j]);
+			store += (size_t)(scenariodata.out_data_size[i][j]);
 	len_to_jump += store;
 
 	if (!SetFPos(file_id, len_to_jump, std::ios_base::beg))
@@ -1007,7 +1007,7 @@ void load_outdoors(
 	{
 		for (short i = 0; i < 9; i++)
 		{
-			file_read_string(file_id, (long)outdoors[targ_x][targ_y].strlens[i], data_store4.outdoor_text[targ_x][targ_y].out_strs[i]);
+			file_read_string(file_id, (size_t)outdoors[targ_x][targ_y].strlens[i], data_store4.outdoor_text[targ_x][targ_y].out_strs[i]);
 		}
 	}
 
@@ -1015,7 +1015,7 @@ void load_outdoors(
 	{
 		for (short i = 0; i < 120; i++)
 		{
-			const long len = (long)(dummy_out.strlens[i]);
+			const size_t len = (size_t)(dummy_out.strlens[i]);
 			if (i == extra)
 			{
 				file_read_string(file_id, len, str);
@@ -1342,7 +1342,7 @@ Boolean load_scenario_header(std::string_view filename, short header_entry)
 		return FALSE;
 
 	for (i = 0; i < 3; i++) {
-		file_read_string(file_id, (long)(short)scenariodata.scen_str_len[i], load_str);
+		file_read_string(file_id, (size_t)(short)scenariodata.scen_str_len[i], load_str);
 		if (i == 0)
 			load_str[29] = 0;
 		else load_str[59] = 0;
